+++ src/termvisage/tui/render.py
@@ -6,7 +6,6 @@ import logging as _logging
 from collections import defaultdict
 from multiprocessing import Event as mp_Event, Lock as mp_Lock, Queue as mp_Queue
 from os import remove
-from os.path import split
 from queue import Empty, Queue
 from tempfile import mkdtemp, mkstemp
 from threading import Event, Lock
@@ -31,14 +30,42 @@ def delete_thumbnail(thumbnail: str) -> bool:
 
 
 def resync_grid_rendering() -> None:
+    # NOTE: The order of operations is very crucial in avoiding deadlocks and even
+    # worse, races. See the resync blocks in `manage_grid_renders()` and
+    # `manage_grid_thumbnails()` especially their beginnings and ends.
+
+    # Signal `GridRenderManager` and `GridThumbnailManager` to **start** resync.
+    #
+    # `GridThumbnailManager` waits for `GridRenderManager` to **start** resync
+    # **before** it does because it modifies (clears) shared data. Hence,
+    # `grid_renderer_in_sync` must be cleared **before** `grid_thumbnailer_in_sync`.
+    grid_renderer_in_sync.clear()
     if main.THUMBNAIL:
-        grid_thumbnail_queue.put(None)  # Send the grid delimiter
         grid_thumbnailer_in_sync.clear()
+
+    # Wait for `GridRenderManager` and `GridThumbnailManager` to **start** resync.
+    #
+    # The order within this set of operations is not necessarily important. However,
+    # the order of this set important with respect to the other sets.
+    grid_renderer_in_sync.wait()
+    if main.THUMBNAIL:
         grid_thumbnailer_in_sync.wait()
 
-    grid_render_queue.put(None)  # Send the grid delimiter
-    grid_renderer_in_sync.clear()
-    grid_renderer_in_sync.wait()
+    # Send the batch delimiter, without which each thread cannot **end** resync.
+    #
+    # It is important that `GridThumbnailManager` **starts** resync **before** sending
+    # the delimiter to `GridRenderManager`, in order to ensure `GridThumbnailManager`
+    # doesn't forward any jobs from the **old** batch to `GridRenderManager` **after**
+    # the delimiter.
+    #
+    # It is also equally important that the delimiter is sent to `GridRenderManager`
+    # **before** `GridThumbnailManager` **ends** resync (which it can't do until it
+    # gets its own delimiter), in order to ensure `GridThumbnailManager` doesn't
+    # forward any jobs from the **new** batch to `GridRenderManager` **before** the
+    # delimiter.
+    grid_render_queue.put(None)
+    if main.THUMBNAIL:
+        grid_thumbnail_queue.put(None)
 
 
 def generate_grid_thumbnails(
@@ -398,7 +425,8 @@ def manage_grid_renders(n_renderers: int):
     subprocesses to render the cells and handles their proper termination.
     Otherwise, it starts a single new thread to render the cells.
     """
-    from . import main
+    from os.path import basename
+
     from .main import ImageClass, grid_active, quitting, update_screen
     from .widgets import Image, ImageCanvas, image_grid
 
@@ -444,26 +472,35 @@ def manage_grid_renders(n_renderers: int):
     for renderer in renderers:
         renderer.start()
 
-    grid_image_size = grid_path = None  # Silence flake8's F821
+    canvas_size = None  # Silence flake8's F821
     faulty_image = Image._ti_faulty_image
-    delimited = False
     grid_cache = Image._ti_grid_cache
     in_sync = grid_renderer_in_sync
+
+    # Since waiting for all renderers to finish all active renders during grid
+    # render syncs is too costly, this is used to filter out any images that get
+    # caught up in a renderer during(/across) grid render sync(s).
+    #
+    # Since it doesn't rotate around until 1000, it's *practically* impossible for an
+    # image to remain in a renderer through that many syncs (by whichever means).
+    grid_batch_no = 0
 
     try:
         while True:
             while not (
-                grid_active.wait(0.1)
+                quitting.is_set()
+                or grid_active.wait(0.1)
                 or quitting.is_set()
                 or not grid_render_out.empty()
             ):
                 pass
+
             if quitting.is_set():
                 break
 
-            if delimited or not in_sync.is_set():
+            if not in_sync.is_set():
                 grid_cache.clear()
-                in_sync.set()
+                in_sync.set()  # Signal "starting resync"
 
                 # Purge the in and out queues and update the loading indicator counter
                 for q in (grid_render_in, grid_render_out):
@@ -475,19 +512,16 @@ def manage_grid_renders(n_renderers: int):
                         else:
                             notify.stop_loading()
 
-                if not delimited:
-                    # Purge all items until the grid delimiter
-                    while grid_render_queue.get():
-                        pass
-                else:
-                    delimited = False
+                # Purge all items up **to** the batch delimiter
+                while grid_render_queue.get():
+                    pass
 
-                grid_path = main.grid_path
+                grid_batch_no = (grid_batch_no + 1) % 1000
                 grid_cell_width = image_grid.cell_width
-                grid_image_size = (grid_cell_width - 2, grid_cell_width // 2 - 2)
+                canvas_size = (grid_cell_width - 2, grid_cell_width // 2 - 2)
                 del grid_cell_width
 
-            if not in_sync.is_set():
+            if quitting.is_set() or not in_sync.is_set():
                 continue
 
             if grid_active.is_set():
@@ -496,32 +530,27 @@ def manage_grid_renders(n_renderers: int):
                 except Empty:
                     pass
                 else:
-                    if not source_and_thumbnail:
-                        delimited = True
-                        continue
-                    grid_render_in.put((*source_and_thumbnail, grid_image_size))
+                    grid_render_in.put(
+                        (grid_batch_no, *source_and_thumbnail, canvas_size)
+                    )
                     notify.start_loading()
 
-            if not in_sync.is_set():
+            if quitting.is_set() or not in_sync.is_set():
                 continue
 
             try:
-                source, thumbnail, render, canvas_size, rendered_size = (
+                batch_no, source, thumbnail, render, rendered_size = (
                     grid_render_out.get(timeout=0.02)
                 )
             except Empty:
                 pass
             else:
-                source_dirname, source_basename = split(source)
                 if (
-                    in_sync.is_set()
-                    # The directory and cell-width checks are to filter out any
-                    # remnants that were caught up in the renderer(s) while syncing
-                    # grid rendering.
-                    and source_dirname == grid_path
-                    and canvas_size == grid_image_size
+                    batch_no == grid_batch_no
+                    and in_sync.is_set()
+                    and not quitting.is_set()
                 ):
-                    grid_cache[source_basename] = (
+                    grid_cache[basename(source)] = (
                         ImageCanvas(
                             render.encode().split(b"\n"), canvas_size, rendered_size
                         )
@@ -540,7 +569,7 @@ def manage_grid_renders(n_renderers: int):
     finally:
         clear_queue(grid_render_in)
         for renderer in renderers:
-            grid_render_in.put((None,) * 3)
+            grid_render_in.put((None,) * 4)
         for renderer in renderers:
             renderer.join()
         clear_queue(grid_render_queue)
@@ -635,7 +664,6 @@ def manage_grid_thumbnails(thumbnail_size: int) -> None:
     generator.start()
     not_generating.set()
 
-    delimited = False
     in_sync = grid_thumbnailer_in_sync
     renderer_in_sync = grid_renderer_in_sync
     thumbnails_to_be_deleted: set[str] = set()
@@ -643,22 +671,29 @@ def manage_grid_thumbnails(thumbnail_size: int) -> None:
     try:
         while True:
             while not (
-                grid_active.wait(0.1)
+                quitting.is_set()
+                or grid_active.wait(0.1)
                 or quitting.is_set()
                 or not thumbnail_out.empty()
                 or thumbnails_to_be_deleted
             ):
                 pass
+
             if quitting.is_set():
                 break
 
-            if delimited or not in_sync.is_set():
-                in_sync.set()
+            if not in_sync.is_set():
+                in_sync.set()  # Signal "starting resync"
+
+                # Wait for `GridRenderManager` to **start** resync before modfying
+                # shared data
                 renderer_in_sync.wait()
 
-                with thumbnail_render_lock:
-                    extra_thumbnail_cache.clear()
-                    thumbnails_being_rendered.clear()
+                # `thumbnail_render_lock` is unnecessary here since `GridRenderManager`
+                # will not access these until `GridThumbnailManager` (this thread) ends
+                # resync and starts forwarding jobs from the new batch over.
+                extra_thumbnail_cache.clear()
+                thumbnails_being_rendered.clear()
 
                 # Purge the in queue and update the loading indicator counter
                 while True:
@@ -695,13 +730,11 @@ def manage_grid_thumbnails(thumbnail_size: int) -> None:
                     delete_thumbnail(thumbnail)
                 thumbnails_to_be_deleted.clear()
 
-                if not delimited:
-                    while grid_thumbnail_queue.get():
-                        pass
-                else:
-                    delimited = False
+                # Purge all items up **to** the batch delimiter
+                while grid_thumbnail_queue.get():
+                    pass
 
-            if not in_sync.is_set():
+            if quitting.is_set() or not in_sync.is_set():
                 continue
 
             if thumbnails_to_be_deleted:
@@ -716,29 +749,24 @@ def manage_grid_thumbnails(thumbnail_size: int) -> None:
                         delete_thumbnail(thumbnail)
                 thumbnails_to_be_deleted -= thumbnails_to_delete
 
-            if not in_sync.is_set():
+            if quitting.is_set() or not in_sync.is_set():
                 continue
 
             if grid_active.is_set():
                 try:
-                    source_and_thumbnail = grid_thumbnail_queue.get(timeout=0.02)
+                    source = grid_thumbnail_queue.get(timeout=0.02)
                 except Empty:
                     pass
                 else:
-                    if not source_and_thumbnail:
-                        delimited = True
-                        continue
-                    if thumbnail := thumbnail_cache.get(
-                        source := source_and_thumbnail[0]
-                    ):
-                        grid_render_queue.put((source, thumbnail))
+                    if thumbnail := thumbnail_cache.get(source):
                         with thumbnail_render_lock:
                             thumbnails_being_rendered[thumbnail].add(source)
+                        grid_render_queue.put((source, thumbnail))
                     else:
                         thumbnail_in.put(source)
                         notify.start_loading()
 
-            if not in_sync.is_set():
+            if quitting.is_set() or not in_sync.is_set():
                 continue
 
             try:
@@ -746,11 +774,11 @@ def manage_grid_thumbnails(thumbnail_size: int) -> None:
             except Empty:
                 pass
             else:
-                if in_sync.is_set():
-                    grid_render_queue.put((source, thumbnail))
+                if in_sync.is_set() and not quitting.is_set():
                     if thumbnail:
                         with thumbnail_render_lock:
                             thumbnails_being_rendered[thumbnail].add(source)
+                    grid_render_queue.put((source, thumbnail))
                 if thumbnail:
                     cache_thumbnail(source, thumbnail, deduplicated)
                 notify.stop_loading()
@@ -858,7 +886,7 @@ def render_grid_images(
     Intended to be executed in a subprocess or thread.
     """
     while True:
-        source, thumbnail, canvas_size = input.get()
+        batch_no, source, thumbnail, canvas_size = input.get()
 
         if not source:  # Quitting
             break
@@ -874,15 +902,15 @@ def render_grid_images(
             image.set_size(Size.FIT if thumbnail else Size.AUTO, maxsize=canvas_size)
             output.put(
                 (
+                    batch_no,
                     source,
                     thumbnail,
                     f"{image:1.1{alpha}{style_spec}}",
-                    canvas_size,
                     image.rendered_size,
                 )
             )
         except Exception:
-            output.put((source, thumbnail, None, canvas_size, None))
+            output.put((batch_no, source, thumbnail, None, None))
 
     clear_queue(output)
 
@@ -939,18 +967,22 @@ extra_thumbnail_cache: dict[str, str] = {}
 # Each value contains the sources currently using the thumbnail
 thumbnails_being_rendered: defaultdict[str, set] = defaultdict(set)
 
-# `GridRenderManager` is actually "in sync" initially.
+# `GridRenderManager` and `GridThumbnailManager` are [taken to be] "in sync" at init.
 #
-# Removing these may result in ocassional deadlocks because at init, the thread
-# may detect "out of sync" and try to prep for a new grid but **without a grid
-# delimiter**.
-# The deadlock is unpredictable and timing-dependent, as it only happens when
-# `main.display_images()` signals "out of sync" **after** the thread has responded to
-# the false initial "out of sync". Hence, the thread blocks on trying to get a grid
-# delimiter.
-# When another "out of sync" is signaled (with a grid delimiter), the thread takes
-# the delimiter as for the previous "out of sync" and comes back around to block
-# again since the event will be unset after consuming the delimiter.
+# Otherwise, it may result in ocassional unpredicatable and timing-sensitive deadlocks
+# at init because either or both of the grid manager threads may read a **false**
+# "out of sync" immediately the grid becomes active and go ahead to **start** a resync.
+#
+# The deadlocks happen when `MainThread` (in `.display_images()`) signals "out of sync"
+# **after** the thread(s) has responded to the **false** initial "out of sync". Hence,
+# the thread(s) blocks, waiting for a batch delimiter in order to **end** the resync.
+#
+# When the **real** initial "out of sync" is signaled, `MainThread` waits for both
+# threads to **[re-]start** resync **before** sending any batch delimiter but since
+# the thread(s) is blocking in wait for a batch delimiter, `MainThread` and both
+# grid manager threads (even it it's only one that read the **false** "out of sync",
+# the other still won't get a batch delimiter after **starting** a resync in response
+# to the **real** "out of sync") simply block forever.
 grid_renderer_in_sync.set()
 grid_thumbnailer_in_sync.set()